tr.centralize now handles std==0 + moved the layers to their own file to keep functiony.py clean
